+++ Project/AcceptanceTests/DataObjects/Guest.java
@@ -1,4 +1,4 @@
-package Mocks;
+package DataObjects;
 
 public class Guest extends User {
