+++ Project/AcceptanceTests/DataObjects/ProductFilter.java
@@ -1,4 +1,4 @@
-package Mocks;
+package DataObjects;
 
 public interface ProductFilter {
     boolean filter(ProductInShop product);
