+++ Project/AcceptanceTests/DataObjects/RegistrationInfo.java
@@ -1,4 +1,4 @@
-package Mocks;
+package DataObjects;
 
 public class RegistrationInfo {
     public String username;
