+++ Project/AcceptanceTests/DataObjects/ShopFilter.java
@@ -1,4 +1,4 @@
-package Mocks;
+package DataObjects;
 
 public interface ShopFilter {
     boolean filter(Shop shop);
