+++ Project/AcceptanceTests/DataObjects/SubscribedUser.java
@@ -1,4 +1,4 @@
-package Mocks;
+package DataObjects;
 
 import java.util.HashMap;
 import java.util.List;
