+++ Project/AcceptanceTests/DataObjects/User.java
@@ -1,4 +1,4 @@
-package Mocks;
+package DataObjects;
 
 import java.util.List;
