+++ Project/AcceptanceTests/Tests/UserTests.java
@@ -3,7 +3,7 @@ package Tests;
 import Bridge.UserProxy;
 import org.junit.jupiter.api.Test;
 
-import Mocks.*;
+import DataObjects.*;
 import Threads.ACEFounderPurchase;
 import Threads.MegaSportFounderPurchase;
 import org.testng.annotations.BeforeClass;
